Documentation files renamed.
