+++ src/doc/cheatsheet.ts
