+++ src/doc/examples.ts
