+++ src/doc/legal.ts
